agents/src/bud2/clients/bud2/generated/pyproject.toml
[tool.poetry]
name = "bud2-api-client"
version = "0.2.0"
description = "A client library for accessing bud2 API"
authors = []
readme = "README.md"
packages = [
    { include = "bud2_api_client" },
]
include = ["bud2_api_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
